requirements.txt
# Core ML Libraries
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
datasets>=2.14.0
tokenizers>=0.15.0

# Parameter-Efficient Fine-tuning
peft>=0.7.0
accelerate>=0.24.0

# Apple Silicon Optimization
mlx>=0.0.9
mlx-lm>=0.0.6

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Evaluation Metrics
evaluate>=0.4.0
rouge-score>=0.1.2
sacrebleu>=2.3.0
bert-score>=0.3.13

# Experiment Tracking and Visualization
wandb>=0.16.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Configuration and Utilities
pyyaml>=6.0
tqdm>=4.65.0
rich>=13.6.0
click>=8.1.0

# Development Tools
jupyter>=1.0.0
ipywidgets>=8.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Optional: Additional evaluation tools
nltk>=3.8.0
spacy>=3.7.0

# Memory profiling (for debugging)
memory-profiler>=0.61.0
psutil>=5.9.0